409-sidebarviews: made sidebarviews dumb (#416)
